Improved Uni-Sign model for WLASL dataset with ISLR task

+++ deformable_attention_2d.py
@@ -123,7 +123,7 @@ class PositionEmbeddingRandom(nn.Module):
         """Positionally encode points that are normalized to [0,1]."""
         # assuming coords are in [0, 1]^2 square and have d_1 x ... x d_n x 2 shape
         coords = 2 * coords - 1
-        coords = coords @ self.positional_encoding_gaussian_matrix
+        coords = coords.float() @ self.positional_encoding_gaussian_matrix
         coords = 2 * np.pi * coords
         # outputs d_1 x ... x d_n x C shape
         return torch.cat([torch.sin(coords), torch.cos(coords)], dim=-1)
@@ -262,7 +262,7 @@ class DeformableAttention2D(nn.Module):
         # pose_init --> [0, 1]
         grid = pose_init[:,None].repeat(1, self.offset_groups, 1, 1) * 2 - 1
         # grid --> [-1, 1]
-        grid = grid.reshape(-1, *pose_init.shape[-2:])
+        grid = grid.reshape(-1, *pose_init.shape[-2:]).float()
 
         vgrid = grid + offsets
         vgrid_scaled = reshape_grid_1d(vgrid)[:,None]
